requirements.txt
argparse==1.4.0
pillow==11.1.0
requests==2.32.3
bs4==0.0.2
torch==2.6.0
numpy==2.2.2
perlin_noise==1.13